refactor: move `bentErrorFixer` function into `src/utils/bent.ts` file
